fixing problems related to bigint

+++ packages/mountainsort2/src/p_extract_segment_timeseries.h
@@ -9,7 +9,7 @@
 #include <QString>
 #include "mlcommon.h"
 
-bool p_extract_segment_timeseries(QString timeseries, QString timeseries_out, int t1, int t2);
-bool p_extract_segment_timeseries_from_concat_list(QStringList timeseries_list, QString timeseries_out, int t1, int t2);
+bool p_extract_segment_timeseries(QString timeseries, QString timeseries_out, bigint t1, bigint t2);
+bool p_extract_segment_timeseries_from_concat_list(QStringList timeseries_list, QString timeseries_out, bigint t1, bigint t2);
 
 #endif // P_EXTRACT_SEGMENT_TIMESERIES_H
